+++ app/helpers/subjects_helper.rb
@@ -1,8 +1,7 @@
 module SubjectsHelper
   def status_helper(subject, long = false)
-    codes = { 'valid' => 'success', 'test' => 'info' }
     content_tag(  :span, (long ? subject.status : subject.status[0]),
-                  class: "label label-#{codes[subject.status.to_s] || 'warning'}",
+                  class: "label label-#{subject.status == 'valid' ? 'success' : 'info'}",
                   rel: 'tooltip',
                   data: { placement: 'left' },
                   title: "#{subject.status unless long}" )