+++ app/models/site.rb
@@ -19,9 +19,4 @@ class Site < ActiveRecord::Base
   has_many :site_users
   has_many :users, -> { where( deleted: false ).order( 'last_name, first_name' ) }, through: :site_users
 
-  # Model Methods
-  def valid_subject_code?(subject_code)
-    subject_code <= "#{self.prefix}#{self.code_maximum}" && subject_code.size <= "#{self.prefix}#{self.code_maximum}".size && subject_code >= "#{self.prefix}#{self.code_minimum}" && subject_code.size >= "#{self.prefix}#{self.code_minimum}".size
-  end
-
 end