+++ db/migrate/20131114195133_change_default_status_to_valid_for_subjects.rb
@@ -1,0 +1,10 @@
+class ChangeDefaultStatusToValidForSubjects < ActiveRecord::Migration
+  def up
+    change_column :subjects, :status, :string, null: false, default: 'valid'
+    Subject.where( status: 'pending' ).update_all( status: 'valid' )
+  end
+
+  def down
+    change_column :subjects, :status, :string, null: false, default: 'pending'
+  end
+end